split the absolute abomination of a class called CustomTriggers.cs into multiple files and changed their namespace
